AmortizeAndWorstCase02: Add a counter to enable learning proper decision trees
When the counter is 0, do not reset. Otherwise, when the counter is 1 or 2 or 3 or etc., do reset.

By contrast, R is not a good candidate as features:
- When R is 0, do not reset.
- When R is not 0 (such as 2181), do reset.
Then, the decision tree will split based on a value that overfits the concrete trace (such as 2181/2), as opposed to 0 (which is more general).